Fix rom loading so that BIOS actually correct detects them and plays them
I had incorrectly assumed that the "g GCE" string followed by the name of the game, etc. were part of a rom header for Vectrex emulators. In fact, the Vectrex BIOS expects this header in the cart rom and reads it to first detect that a cartridge is inserted (otherwise it plays the internal Minestorm), and then reads the title and music to display and play.