Fix dots not drawing at all
Problem was that I assumed the beam would only draw if /RAMP (integrators) was enabled, but this is not true. Integrators control movement of the beam, while beam on/off is independent. Also needed to detect when we were attempting to draw a dot and draw GL_POINTS rather than GL_LINES since drawing a line from a point to itself doesn't draw anything in OpenGL.

Also opportunistically added multiply and divide by scalar for Vector2 and improved the delta movement computation code to take advantage of it.