Clean up mess around passing screen vs window size to GLRender
Now we only pass in the window size to SetViewport, and resources that depend on window size are (re)created in that function (i.e. texture targets). The screen size constants, renamed to VECTREX_SCREEN_WIDTH/HEIGHT, have been moved to GLRender.cpp and are used to compute the projection matrix in SetViewport.