Restore initial current working directory after opening file
It's useful to maintain the initial CWD so that output files (traces, imgui.ini) end up in the right place.
We also remember the path of the last loaded rom file so that we can set CWD to it's directory before opening the file dialog so that we browse in the same directory as the last time.