Cpu: implement CWAI along with IRQ handling
CWAI is a bit weird because it's the first (and only?) instruction we implement that doesn't fully "execute" from the perspective of the Cpu::ExecuteInstruction() call. As such, the debugger has to do a little more work to detect whether ExecuteInstruction actually did execute something (if cycles returned > 0), and make sure not to add trace, etc. if no instruction really executed.