Undo adding a screen border in GLRender and instead apply a global line draw scale in Screen
This allows us to stick with the 256x256 grid size when rendering, which makes way more sense. The fact that Screen adds delays is an internal detail, but it should continue to produce lines that fit within the expected grid.
Along with this change, we undo render scale tweaks we made earlier as we no longer need to compensate for the larger grid render.