Fix direct samples not sounding right when both direct samples and PSG are used at the same time to produce audio
In the end, saving just the last direct sample, and attempting to clear it when writing to PSG just can't work when both are used at the same time. Instead, we simply accumulate all direct audio samples, then produce the average value when we need to produce an output sample.