Add "clock sync" for devices so that devices (VIA) update on read/write for more cycle-accurate emulation
Fixes the Bedlam lockup bug after when starting a second new game. The Bedlam bug happens because of the following:

- Wait_Recal gets called, it starts looping waiting for T2 enable flag.
- We execute "bita", which stores that T2 is not enabled, but then we update Via, which enables it, then when we are about to execute 'beq', we end up triggering the IRQ handler; the IRQ handler then runs and eventually disables T2 interrupt. When it returns (RTI), the 'beq' will jump back to the bita because T2 wasn't enabled, and we start all over, and end up in an infinite loop.

1. bita    <VIA_int_flags --> T2 disabled so CC.Z high
2. update VIA -> T2 enabled
3. beq     LF19E --> don't execute, instead we execute IRQ
4. IRQ runs, T2 disabled, RTI
5. beq     LF19E --> execute for real this time, CC.Z is still high (popped off stack), we jump to 'bita' and start over again (loop forever)

This commit fixes the issue by making it so that when 'bita' executes, it will sync VIA upon reading from it, which ends up raising the VIA interrupt flag, and this is seen by 'bita', so CC.Z will be low. After that, even though the IRQ handler runs, when it returns, 'beq' will not branch, and thus not start up the loop again.