FPS: don't correct for overtime, and display integer fps
Using overflowed time for the next frame's time after computing fps doesn't really work since we don't adjust the frame count itself, so this results in a reported fps that's consistently below 60. Instead, we just drop the overflow time, and display a rounded FPS, and with vsync, we are at a fairly solid 60 fps, as expected.