DevMemoryDevice: fix printing of multiple strings
When attempting to printf multiple strings
(e.g. vectrexy::printf("%s %s, a, b)), on the second emplace_back into
the strings vector, the heap buffer would be relocated, invalidating the
string addresses stored in the args byte buffer. We now resever enough
space to store 1024 strings, which is way more than enough, to ensure
that the vector never gets relocated.

Also added a "[DEV] " prefix to this output so that we can tell that it
comes from the dev device.