Debugger: support abnormal call stack returns and recursion
Call stack tracking now also detects when functions return abnormally;
that is, when the return address is manually popped and discarded off
the stack, without setting PC to its value, as Bedlam does. These are
now detected, frames popped off the stack correctly, and a warning
emitted.

More with this change:
* Conditional breakpoint support was added, although not exposed to the
user yet. These are used to implement next and finish commands.
* Next and Finish have been modified to track the call stack size,
rather than break on the return address. This allows for these commands
to work as expected on recursive functions.
* Call stack tracking for CWAI has been removed for now, as it does not
work as expected. The problem is that CWAI is currently implemented to
push PC (and all registers) to the stack, but the jump to the target
function happens later, when the next interrupt occurs. Call stack
tracking assumes the call performs the push and jump at once, or bad
things happen. To fix this, we should modify CWAI so that it delays the
register pushing once the interrupt happens, so that it happens in the
same frame as the jump.
* Fixed all SWI calls to retrieve the return address at the correct
offset on the stack (offset 10, rather than 12).
* Refactored step, continue, next, finish, and until in terms of new
local Step and Continue lambdas so that they all behave consistently.