DapDebugger: make StackTraceRequest and SetBreakpointsRequest thread safe
Both these requests modify data members also accessed by the main thread.
We now have a "frame mutex" that is locked once per frame, and is locked
by these requests.

Also fix an infinite loop that would occur while in the Paused state. We
made sure to clear m_cpuCyclesLeft upon entering Paused, but while
remaining in that state, we wouldn't clear it, so we'd end up infinitely
looping in ExecuteFrameInstructions. This would now deadlock with a
frame mutex lock, since we need to leave this scope to release the lock
for the StackTraceRequest and SetBreakpointsRequest.