DapDebugger: fix breakpoints not working due to drive letter case mismatch on Windows
Unfortunately, VS Code isn't always consistent with the case of driver
letter of the SetBreakpointsRequest file path. Microsoft has documented
that they will not fix this (https://github.com/microsoft/vscode/issues/43959),
so we must take care of it on our end. On Windows, we always lower-case
the file path drive letter, if any.