RstParser: refactor regex into classes
Instead of using std::regex directly, wrap it up into a MatchBase class,
from which we derive a class for each regex, along with functions to
name and return each match group. This makes it easier to read the code,
and to be able to replace the regex library in the future.

Also moved out move constants into an anonymous namespace for clarity.