RstParser: fix assert when function has multiple scopes, but no variables declared in the outermost scope
For example, a function like:

```
int main() {
    { int a = 0; }
    { int b = 0; }
    return 0;
}
```

Produces the following stabs output:

```
91 ;	.stabs	"main:F7",36,0,0,_main
92 ;	.stabs	"a:7",128,0,0,0
93 ;	.stabn	192,0,0,LBB2
94 ;	.stabn	224,0,0,LBE2
95 ;	.stabs	"b:7",128,0,0,1
96 ;	.stabn	192,0,0,LBB3
97 ;	.stabn	224,0,0,LBE3
```

The problem here is that there's no root scope info for the function
itself, since the function has no variables delcared at the root scope.
Before this fix, my code assumed there would always be root scope info
in the stabs output, so when the above situation occured, after parsing
the first inner scope, ` { int a = 0; }`, I'd assume the function was
finished being defined. Upon reaching the next scope, ` { int b = 0; }`,
we'd hit an assertion because we no longer were defining a function. The
fix here is to always add a root scope when defining a new function, and
when done defining the function, we remove it if it only has one child.